requirements.txt
ale-py==0.8.1  # 0.7.4 original
einops==0.3.2
# gym 0.21 needs these
setuptools==65.5.0
wheel==0.38.4
# end
gym[accept-rom-license]==0.21.0
hydra-core==1.1.1
opencv-python
pillow==9.4.0
protobuf==3.20.*
psutil==5.8.0
pygame==2.1.2
tqdm==4.62.3
wandb==0.17.0
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.18.1+cu118
boto3==1.26.*
paramiko==3.2.0